feat(FR-1804): enhance user management table with flexible column composition (#4863)
Resolves #4862 ([FR-1804](https://lablup.atlassian.net/browse/FR-1804))

## Background

The current user management table implementation uses a fixed array-based approach for column composition (`extraColumns?: BAIColumnType[]`). This approach has limitations:

1. **Limited Flexibility**: Can only prepend or append columns, cannot insert at specific positions
2. **Tight Coupling**: Query logic and presentation are coupled in a single component
3. **Difficult Maintenance**: Changes to column ordering require modifying the base component

## Purpose

This enhancement introduces a function-based column composition pattern to address these limitations:

1. **Architectural Separation**: Separate data fetching (query orchestrator) from presentation (fragment-based table component)
2. **Flexible Composition**: Enable inserting, filtering, reordering columns at any position
3. **Better Developer Experience**: Provide render-props pattern for maximum control over column composition
4. **Maintainability**: Allow parent components to customize columns without modifying base component

## Benefits

- **Flexibility**: Insert control columns at specific positions (e.g., after email column)
- **Reusability**: Base table component can be reused with different column compositions
- **Type Safety**: TypeScript ensures column customization is type-safe
- **Clean Architecture**: Follows separation of concerns principle (data fetching vs presentation)

## Implementation Details

### Changed API

**Before:**
```typescript
extraColumns?: BAIColumnType<UserNodeInList>[]
```

**After:**
```typescript
customizeColumns?: (baseColumns: BAIColumnType<UserNodeInList>[]) => BAIColumnType<UserNodeInList>[]
```

### Usage Example

```typescript
<UserNodes
  usersFrgmt={users}
  customizeColumns={(baseColumns) => [
    baseColumns[0],          // Keep email column first
    controlColumn,           // Insert control column
    ...baseColumns.slice(1)  // Rest of base columns
  ]}
/>
```

## Changes

- New `UserNodes.tsx` component with fragment-based GraphQL pattern
- Function-based `customizeColumns` prop for flexible composition
- Updated `UserManagement.tsx` to use new composition pattern
- Added translations for 21 languages
- Removed `UserNodeList.tsx` (replaced by UserNodes)

## Test Plan

**Checklist:**

- [ ] Verify user list displays correctly with all columns
- [ ] Check control column is inserted at correct position (after email)
- [ ] Test column sorting functionality
- [ ] Verify status filter (active/inactive) works
- [ ] Test search and filtering
- [ ] Confirm user actions (edit, enable/disable) work correctly
- [ ] Check all 21 language translations display properly

[FR-1804]: https://lablup.atlassian.net/browse/FR-1804?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
