+++ src/lib/backend.ai-client-esm.ts
@@ -825,6 +825,9 @@ class Client {
     if (this.isManagerVersionCompatibleWith('25.15.0')) {
       this._features['agent-stats'] = true;
     }
+    if (this.isManagerVersionCompatibleWith('25.15.2')) {
+      this._features['user-node-query-project-filter'] = true;
+    }
     if (this.isManagerVersionCompatibleWith('25.16.0')) {
       this._features['multi-agents'] = true;
     }
